+++ ChoGGi's Library/Changelog.md
@@ -2,6 +2,7 @@
 
 ## v6.8 (Unreleased)
 - Export Map Data NESW strings were missing.
+- Colour Modifier checkmarks were broked.
 
 ## v6.7 (May 30 2019)
 - Cernan.